Update Controller to use Service Part 2

+++ src/DTOs/CauHoi/CauHoiDTO.cs
@@ -8,5 +8,5 @@ public class CauHoiDTO
     public string Ten { get; set; } = string.Empty;
     public decimal TrongSo { get; set; }
     public int BaiKiemTraId { get; set; }
-    public int ThangDiem { get; set; }
+    public decimal ThangDiem { get; set; }
 }
